Add download link tests, move shared helpers to util folder
