+++ tests/playwright/specs/download-attribution/remove.spec.js
@@ -8,117 +8,10 @@
 
 const { test, expect } = require('@playwright/test');
 const openPage = require('../../scripts/open-page');
-
-/**
- * Sets up a route mock for stub_attribution_code XHR requests.
- *
- * Captured query params are written to capturedStubParams
- * (passed by reference via wrapper).
- *
- * The mock responds with a signed attribution payload so downstream cookie
- * writes complete normally.
- * @param {import('@playwright/test').Page} page
- * @param {{ params: Object | null }} capture - Object whose .params property
- *   is set to the captured URLSearchParams entries when the request fires.
- */
-async function routeStubAttributionCode(page, capture) {
-    await page.route('**/stub_attribution_code/**', async (route) => {
-        const reqUrl = new URL(route.request().url());
-        capture.params = Object.fromEntries(reqUrl.searchParams);
-        await route.fulfill({
-            status: 200,
-            contentType: 'application/json',
-            body: JSON.stringify({
-                attribution_code: 'mocked-attribution-code',
-                attribution_sig: 'mocked-attribution-sig'
-            })
-        });
-    });
-}
-
-/**
- * Intercepts Mozilla.DownloadAttribution.getGtagClientID before page scripts
- * run so it returns a fixed GA4 client ID without triggering real GA4 network
- * calls.
- *
- * Pass to page.addInitScript() before navigation.
- */
-function mockGetGtagClientID() {
-    let _daValue;
-    const _mozilla = {};
-
-    Object.defineProperty(window, 'Mozilla', {
-        configurable: true,
-        enumerable: true,
-        get() {
-            return _mozilla;
-        },
-        set(val) {
-            if (val && typeof val === 'object') {
-                Object.assign(_mozilla, val);
-            }
-        }
-    });
-
-    Object.defineProperty(_mozilla, 'DownloadAttribution', {
-        configurable: true,
-        enumerable: true,
-        get() {
-            return _daValue;
-        },
-        set(val) {
-            if (val && typeof val === 'object') {
-                val.getGtagClientID = () => 'mocked-ga4-client-id';
-            }
-            _daValue = val;
-        }
-    });
-}
-
-/**
- * Sets data-stub-attribution-campaign-force on <html> before page scripts run.
- * Pass to page.addInitScript() before navigation.
- */
-function forceEssentialCampaign(campaign = 'smart_window') {
-    function setAttr() {
-        document.documentElement.setAttribute(
-            'data-stub-attribution-campaign-force',
-            campaign
-        );
-    }
-    // Init scripts run before HTML is parsed, so document.documentElement
-    // may be null. Observe the document for the <html> element if needed.
-    if (document.documentElement) {
-        setAttr();
-    } else {
-        new MutationObserver(function (_, obs) {
-            if (document.documentElement) {
-                obs.disconnect();
-                setAttr();
-            }
-        }).observe(document, { childList: true, subtree: true });
-    }
-}
-
-/**
- * Prevents the download-as-default feature from interfering with
- * attribution tests by blocking the JS bundles and removing checkbox
- * elements from the DOM before page scripts run.
- * @param {import('@playwright/test').Page} page
- */
-async function removeDownloadAsDefault(page) {
-    await page.route('**/js/download_as_default*.js', (route) => route.abort());
-    await page.addInitScript(() => {
-        const obs = new MutationObserver(() => {
-            const labels = document.querySelectorAll('.default-browser-label');
-            if (labels.length > 0) {
-                labels.forEach((el) => el.remove());
-                obs.disconnect();
-            }
-        });
-        obs.observe(document, { childList: true, subtree: true });
-    });
-}
+const mockGetGtagClientID = require('./utils/mock-get-gtag-client-id');
+const routeStubAttributionCode = require('./utils/route-stub-attribution-code');
+const forceEssentialCampaign = require('./utils/force-essential-campaign');
+const removeDownloadAsDefault = require('./utils/remove-download-as-default');
 
 const existingAnalyticsParams =
     'utm_source=newsletter&utm_medium=email&utm_campaign=existing';
