+++ tests/playwright/specs/download-attribution/utils/route-stub-attribution-code.js
@@ -1,0 +1,36 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+'use strict';
+
+/**
+ * Sets up a route mock for stub_attribution_code XHR requests.
+ *
+ * Captured query params are written to capturedStubParams
+ * (passed by reference via wrapper).
+ *
+ * The mock responds with a signed attribution payload so downstream cookie
+ * writes complete normally.
+ * @param {import('@playwright/test').Page} page
+ * @param {{ params: Object | null }} capture - Object whose .params property
+ *   is set to the captured URLSearchParams entries when the request fires.
+ */
+async function routeStubAttributionCode(page, capture) {
+    await page.route('**/stub_attribution_code/**', async (route) => {
+        const reqUrl = new URL(route.request().url());
+        capture.params = Object.fromEntries(reqUrl.searchParams);
+        await route.fulfill({
+            status: 200,
+            contentType: 'application/json',
+            body: JSON.stringify({
+                attribution_code: 'mocked-attribution-code',
+                attribution_sig: 'mocked-attribution-sig'
+            })
+        });
+    });
+}
+
+module.exports = routeStubAttributionCode;
